requirements.txt
psutil
lgpio
pillow
numpy
# smbus2
spidev
gpiozero
netifaces
influxdb-client[ciso]